remove SequentialMergeStore in favor of async-lock
