+++ apps/hubble/package.json
@@ -30,16 +30,17 @@
   "devDependencies": {
     "@faker-js/faker": "~7.6.0",
     "@libp2p/interface-mocks": "^9.0.0",
+    "@types/async-lock": "^1.4.0",
     "@types/chance": "^1.1.3",
     "@types/node-cron": "^3.0.7",
     "@types/progress": "^2.0.5",
+    "@types/rwlock": "^5.0.3",
     "chance": "~1.1.9",
     "csv-stringify": "~6.2.3",
     "eslint-config-custom": "*",
     "fishery": "~2.2.2",
     "pino-pretty": "~9.1.1",
     "progress": "~2.0.3",
-    "@types/rwlock": "^5.0.3",
     "ts-mockito": "~2.6.1"
   },
   "dependencies": {
@@ -48,7 +49,6 @@
     "@farcaster/protobufs": "0.1.3",
     "@farcaster/utils": "0.2.2",
     "@grpc/grpc-js": "~1.8.7",
-    "@multiformats/multiaddr": "^11.0.0",
     "@libp2p/interface-connection": "^3.0.2",
     "@libp2p/interface-peer-id": "^2.0.0",
     "@libp2p/mplex": "^7.0.0",
@@ -56,6 +56,8 @@
     "@libp2p/pubsub-peer-discovery": "^8.0.0",
     "@libp2p/tcp": "^6.0.0",
     "@libp2p/utils": "^3.0.2",
+    "@multiformats/multiaddr": "^11.0.0",
+    "async-lock": "^1.4.0",
     "commander": "~10.0.0",
     "ethers": "~5.7.2",
     "libp2p": "0.42.2",
